# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ServiceResponse_any-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {},
						"type": "array",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ServiceResponse_any-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"anyOf": [],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"InterventionAreaCreateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"documents": {
						"items": {
							"properties": {
								"deleteToken": {
									"type": "string",
									"nullable": true
								},
								"publicId": {
									"type": "string",
									"nullable": true
								},
								"userId": {
									"type": "string"
								},
								"documentUrl": {
									"type": "string",
									"nullable": true
								},
								"addedAt": {
									"type": "string",
									"format": "date-time"
								},
								"type": {
									"type": "string",
									"nullable": true
								},
								"size": {
									"type": "number",
									"format": "double",
									"nullable": true
								},
								"documentName": {
									"type": "string"
								}
							},
							"required": [
								"userId",
								"documentName"
							],
							"type": "object"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InterventionAreaUpdateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"documents": {
						"items": {
							"properties": {
								"deleteToken": {
									"type": "string"
								},
								"publicId": {
									"type": "string"
								},
								"userId": {
									"type": "string"
								},
								"documentUrl": {
									"type": "string"
								},
								"addedAt": {
									"type": "string",
									"format": "date-time"
								},
								"type": {
									"type": "string"
								},
								"size": {
									"type": "number",
									"format": "double"
								},
								"documentName": {
									"type": "string"
								}
							},
							"required": [
								"userId",
								"documentName"
							],
							"type": "object"
						},
						"type": "array"
					},
					"removeDocumentIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"DocumentCreateRequest": {
				"properties": {
					"projectId": {
						"type": "string"
					},
					"documentName": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"type": {
						"type": "string",
						"nullable": true
					},
					"addedAt": {
						"type": "string",
						"format": "date-time"
					},
					"documentUrl": {
						"type": "string",
						"nullable": true
					},
					"userId": {
						"type": "string"
					},
					"publicId": {
						"type": "string",
						"nullable": true
					},
					"deleteToken": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"documentName",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"optionalJwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "community-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Ikundabayo placide"
		}
	},
	"paths": {
...
		"/api/intervention-areas": {
			"get": {
				"operationId": "GetInterventionAreas",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Intervention Areas"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateInterventionArea",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Intervention Areas"
				],
				"security": [
					{
						"jwt": [
							"intervention_area:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InterventionAreaCreateRequest"
							}
						}
					}
				}
			}
		},
		"/api/intervention-areas/{interventionAreaId}": {
			"get": {
				"operationId": "GetInterventionAreaById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					},
					"404": {
						"description": "Intervention area not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Intervention Areas"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "interventionAreaId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateInterventionArea",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Intervention Areas"
				],
				"security": [
					{
						"jwt": [
							"intervention_area:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "interventionAreaId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InterventionAreaUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteInterventionArea",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Intervention Areas"
				],
				"security": [
					{
						"jwt": [
							"intervention_area:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "interventionAreaId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/documents": {
			"get": {
				"operationId": "GetDocuments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Documents"
				],
				"security": [
					{
						"jwt": [
							"project:read"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "projectId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateDocument",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Documents"
				],
				"security": [
					{
						"jwt": [
							"project:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DocumentCreateRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}